// frontend/src/app/services/models/registration-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Location } from '../models/location';
export interface RegistrationRequest {
  email: string;
  experience?: number;
  farmLocation?: Location;
  farmType?: Array<string>;
  fullName: string;
  password: string;
  phoneNumber: string;
}